Adding build rules to enable more options in bin/mono
