+++ src/extra_data_library/build.sh
@@ -1,4 +1,5 @@
 LIB_STATIC_LIBS="config_helpers utils tinyexpr"
 
-COMPILE_SHARED_LIB "default_extra_data" "extra_data.c" "" "${LIB_STATIC_LIBS}"
-COMPILE_SHARED_LIB "mixed_models_extra_data" "extra_mixed_models.c" "" "${LIB_STATIC_LIBS}"
+CHECK_CUSTOM_FILE "custom_extra_data_functions.c"
+
+COMPILE_SHARED_LIB "default_extra_data" "extra_data.c ${CUSTOM_FILE}" "" "${LIB_STATIC_LIBS}"
