+++ src/linear_system_solver_library/build.sh
@@ -2,4 +2,6 @@ if [ -n "$CUDA_FOUND" ]; then
     EXTRA_CUDA_LIBS="cudart cublas cusparse"
 fi
 
-COMPILE_SHARED_LIB "default_linear_system_solver" "linear_system_solver.c" "" "alg config_helpers utils tinyexpr gpu_utils" "$EXTRA_CUDA_LIBS" "$CUDA_LIBRARY_PATH"
+CHECK_CUSTOM_FILE "custom_linear_system_solver_functions.c"
+
+COMPILE_SHARED_LIB "default_linear_system_solver" "linear_system_solver.c ${CUSTOM_FILE}" "" "alg config_helpers utils tinyexpr gpu_utils" "$EXTRA_CUDA_LIBS" "$CUDA_LIBRARY_PATH"
