+++ src/matrix_assembly_library/build.sh
@@ -1,6 +1,8 @@
 LIB_STATIC_DEPS="utils alg config_helpers tinyexpr sds"
 
-COMPILE_SHARED_LIB "default_matrix_assembly" "matrix_assembly.c" "" "${LIB_STATIC_DEPS}"
+CHECK_CUSTOM_FILE "custom_matrix_assembly_functions.c"
+
+COMPILE_SHARED_LIB "default_matrix_assembly" "matrix_assembly.c ${CUSTOM_FILE}" "" "${LIB_STATIC_DEPS}"
 
 if [ -n "$COMPILE_WITH_DDM" ]; then
     COMPILE_SHARED_LIB "ddm_matrix_assembly" "ddm_matrix_assembly.c" "" "${LIB_STATIC_DEPS}"
