+++ src/postprocessing_library/build.sh
@@ -1,1 +1,3 @@
-COMPILE_SHARED_LIB "default_postprocess" "postprocessing.c" "" "config_helpers utils tinyexpr vtk_utils yxml miniz" "" ""
+CHECK_CUSTOM_FILE "custom_postprocessing_functions.c"
+
+COMPILE_SHARED_LIB "default_postprocess" "postprocessing.c ${CUSTOM_FILE}" "" "config_helpers utils tinyexpr vtk_utils yxml miniz" "" ""
