+++ src/purkinje_library/build.sh
@@ -1,1 +1,3 @@
-COMPILE_SHARED_LIB "default_purkinje" "purkinje.c purkinje_helpers.c" "purkinje_helpers.h" "alg config_helpers utils solvers graph tinyexpr"
+CHECK_CUSTOM_FILE "custom_purkinje_functions.c"
+
+COMPILE_SHARED_LIB "default_purkinje" "purkinje.c ${CUSTOM_FILE} purkinje_helpers.c" "purkinje_helpers.h" "alg config_helpers utils solvers graph tinyexpr"
