+++ src/restore_state_library/build.sh
@@ -2,4 +2,6 @@ if [ -n "$CUDA_FOUND" ]; then
     EXTRA_CUDA_LIBS="cudart"
 fi
 
-COMPILE_SHARED_LIB "default_restore_state" "restore_state.c" "" "alg config_helpers utils sds tinyexpr" "$EXTRA_CUDA_LIBS" "$CUDA_LIBRARY_PATH"
+CHECK_CUSTOM_FILE "custom_restore_state_functions.c"
+
+COMPILE_SHARED_LIB "default_restore_state" "restore_state.c ${CUSTOM_FILE}" "" "alg config_helpers utils sds tinyexpr" "$EXTRA_CUDA_LIBS" "$CUDA_LIBRARY_PATH"
