+++ src/stimuli_library/build.sh
@@ -1,1 +1,3 @@
-COMPILE_SHARED_LIB "default_stimuli" "stimuli.c" "" "alg config_helpers utils tinyexpr" "m"
+CHECK_CUSTOM_FILE "custom_stimuli_functions.c"
+
+COMPILE_SHARED_LIB "default_stimuli" "stimuli.c ${CUSTOM_FILE}" "" "alg config_helpers utils tinyexpr" "m"
